+++ src/Movement/MoveSegment.cpp
@@ -45,7 +45,7 @@ void MoveSegment::ReleaseAll(MoveSegment *item) noexcept
 
 void MoveSegment::DebugPrint(char ch) const noexcept
 {
-	debugPrintf("%c s=%" PRIu32 " t=%.1f d=%.2f u=%.4e a=%.4e f=%02" PRIx32 "\n", ch, (uint32_t)startTime, (double)duration, (double)distance, (double)u, (double)a, flags.all);
+	debugPrintf("%c s=%" PRIu32 " t=%" PRIu32 " d=%.2f u=%.4e a=%.4e f=%02" PRIx32 "\n", ch, startTime, duration, (double)distance, (double)u, (double)a, flags.all);
 }
 
 /*static*/ void MoveSegment::DebugPrintList(char ch, const MoveSegment *segs) noexcept
